==> 2-pira-scre.ipynb <==
# Import python libraries
import numpy as np
import pandas as pd
import importlib
import random
from tqdm import tqdm
import datetime
import geopandas as gpd
from concurrent.futures import ThreadPoolExecutor, as_completed

# Set random seed
random.seed(42)
np.random.seed(42)

# import 3 python files that can be reloaded if any changes are made to them
import st_toolkit as geohl
importlib.reload(geohl)

import cri_calc_new as cri
importlib.reload(cri)

import cri_helper as helper
importlib.reload(helper)

# calculates 5 target factors and the cri for any pair of ships
def calculate_cri(rec_own, rec_target):
    own = rec_own._asdict()
    target = rec_target._asdict()

    ves_dcpa, ves_tcpa, hr, rel_movement_angle, dist_euclid, speed_r = cri.colregs_alarms(own=own, target=target)
    ves_cri = cri.calculate_cri(own, target, ves_dcpa, ves_tcpa, hr, rel_movement_angle, dist_euclid, speed_r)

    return ves_dcpa, ves_tcpa, hr, rel_movement_angle, dist_euclid, speed_r, ves_cri

# calculates the 5 target factors for all ships in the dataset with respect to every other ship
# also calculates the cri between them
def calculate_cri_timeslice(df):
    timeslice_result = []

    for row_i in df.itertuples():
        for row_j in df.itertuples():
            if row_i.Index == row_j.Index:
                continue

            timeslice_result.append([row_i.Index, row_i.mmsi, row_i.geom, row_i.speed, row_i.course,
                                     row_j.Index, row_j.mmsi, row_j.geom, row_j.speed, row_j.course, *calculate_cri(row_i, row_j)])

#     return pd.DataFrame(timeslice_result, columns=['own', 'target', 'dcpa', 'tcpa', 'hr', 'rel_movement_angle', 'dist_euclid', 'speed_r', 'cri'])
    return pd.DataFrame(timeslice_result, columns=['own_Index', 'own_mmsi', 'own_geom', 'own_speed', 'own_course',
                                                   'target_Index', 'target_mmsi', 'target_geom', 'target_speed', 'target_course',
                                                   'dcpa', 'tcpa', 'hr', 'rel_movement_angle', 'dist_euclid', 'speed_r', 'cri'])

# reading the pre-processed pickel file from the pre-processing notebook
gdf_sub_moving = pd.read_pickle('Final\\gdf_sub_moving.pickle')

gdf_sub_moving

Calculate CRI

# using parallel processing to calculate the target factors and cri for all own-ship-target-ship pairs
from joblib import Parallel, delayed
import multiprocessing


def applyParallel(dfGrouped, func, n_jobs=-1):
    n_jobs = multiprocessing.cpu_count() if n_jobs == -1 else n_jobs
    print(f'Scaling {func} to {n_jobs} CPUs')

    retLst = Parallel(n_jobs=n_jobs)(delayed(func)(group) for name, group in tqdm(dfGrouped))
    return pd.concat(retLst)


gdf_scre = applyParallel(gdf_sub_moving.groupby(['datetime']), calculate_cri_timeslice, 18)

gdf_scre

# merge the dfs based on own index
gdf_scre_mi = gdf_scre.copy()
gdf_scre_mi = pd.merge(gdf_scre_mi, gdf_sub_moving.datetime, left_on='own_Index', right_index=True)
gdf_scre_mi.set_index(['datetime'], inplace=True)

gdf_scre_mi

date_times = gdf_scre_mi.index

# Convert to Unix timestamp
gdf_scre_mi['timestamp_'] = (pd.to_datetime(date_times) - pd.Timestamp("1970-01-01")) / pd.Timedelta('1s')

gdf_scre_mi

saint_input = gdf_scre_mi.loc[gdf_scre_mi.own_Index.isin(gdf_sub_moving.index.values)].copy()

saint_input

saint_input.to_pickle('Final\\processed_data.pickle')

# input and ground truth values of model
X = saint_input[['dist_euclid', 'own_speed', 'target_speed', 'own_course', 'target_course']].values
y = saint_input[['cri']].values.ravel()

# convert to pickel file
X_df = pd.DataFrame(X, columns=['dist_euclid', 'own_speed', 'target_speed', 'own_course', 'target_course'])
X_df.to_pickle('Final\\X_df.pkl')

# convert ground truths to pickel file
y_cri_df = pd.DataFrame(y,columns=['cri'])
y_cri_df.to_pickle('Final\\y_cri_df.pkl')
