==> 3-pira-enfactors.ipynb <==
# import all necessary libraries
import numpy as np
import pandas as pd
import importlib
import random
from tqdm import tqdm
import datetime
import geopandas as gpd
from concurrent.futures import ThreadPoolExecutor, as_completed

from pykrige.ok import OrdinaryKriging

import st_toolkit as geohl
importlib.reload(geohl)

# Set random seed
random.seed(42)
np.random.seed(42)

processed_data = pd.read_pickle('Final\\processed_data.pickle')

processed_data

# obtain array of lat and lon of all own ship samples from processed_data
import re
# import numpy as np
from concurrent.futures import ThreadPoolExecutor
# from tqdm import tqdm

loc_own = processed_data['own_geom']
loc_tar = processed_data['target_geom']

# Define a function to process a single location
def process_location(temp_loc):
    point_string = str(temp_loc)
    numbers = re.findall(r"[-+]?\d*\.\d+|\d+", point_string)
    return float(numbers[0]), float(numbers[1])

lat1 =np.array([])
lon1 =np.array([])

# Define the number of threads (adjust as needed)
num_threads = 16

# Create a ThreadPoolExecutor
with ThreadPoolExecutor(max_workers=num_threads) as executor:
    # Process each location in parallel
    results = list(tqdm(executor.map(process_location, loc_own), total=len(loc_own), desc='Creating location datasets'))

# Extract lat and lon from results
lon1, lat1 = zip(*results)

# Convert to numpy arrays
lon1 = np.array(lon1)
lat1 = np.array(lat1)

# Print a success message
print("Parallel processing completed.")

# obtain array of lat and lon of all target ship samples from processed_data
lat2 =np.array([])
lon2 =np.array([])

# Define the number of threads (adjust as needed)
num_threads = 16

# Create a ThreadPoolExecutor
with ThreadPoolExecutor(max_workers=num_threads) as executor:
    # Process each location in parallel
    results = list(tqdm(executor.map(process_location, loc_tar), total=len(loc_tar), desc='Creating location datasets'))

# Extract lat and lon from results
lon2, lat2 = zip(*results)

# Convert to numpy arrays
lon2 = np.array(lon2)
lat2 = np.array(lat2)

# Print a success message
print("Parallel processing completed.")

dfw = pd.read_csv('noaa_weather_jul2018_v2.csv', parse_dates=['timestamp'])
gdfw = geohl.getGeoDataFrame_v2(dfw, crs='epsg:4326')
gdfw2 = gdfw.loc[gdfw.timestamp.between(datetime.datetime(2018, 7, 12, 0, 0, 0), datetime.datetime(2018, 7, 13, 0, 0, 0), inclusive='both')].copy()
gdfw2_cols_dropped = gdfw2.drop(['TMP', 'TMIN', 'TMAX', 'PRMSL', 'RH', 'DPT', 'UGRD', 'VGRD', 'geom', 'WSPD', 'WDIRMAT', 'APCP'], axis = 1)
gdfw2_sorted = gdfw2_cols_dropped.sort_values(by=['lon', 'lat'])


num_rows = len(gdfw2_sorted)

# obtaining the 'subset size'- it is the number of timestamps in a day (including 12am the next day) for any given station
subset_size = 1 # initializing the subset size

for i in range(1, num_rows):
  if (gdfw2_sorted.iloc[i]['lat'] == gdfw2_sorted.iloc[i-1]['lat']) and (gdfw2_sorted.iloc[i]['lon'] == gdfw2_sorted.iloc[i-1]['lon']):
    subset_size += 1
  else:
    break

# print(subset_size)
num_subsets = (num_rows - 1) // subset_size + 1
subset_indices = [int(i * num_rows / num_subsets) for i in range(num_subsets)]
# print(subset_indices)


interpolated_subsets = []
for start_index in subset_indices:
    end_index = start_index + subset_size
    subset = gdfw2_sorted.iloc[start_index:end_index]

    subset = subset.set_index('timestamp').sort_index()
    subset_interpolated = subset.resample('1S').asfreq() # interpolating time into 5 second intervals

    subset_interpolated['GUST'] = subset_interpolated['GUST'].interpolate() #interpolating gust
    subset_interpolated['VIS'] = subset_interpolated['VIS'].interpolate() #interpolating vis
    subset_interpolated['WDIRMET'] = subset_interpolated['WDIRMET'].interpolate() #interpolating mathematical wind direction
    # subset_interpolated['WSPD'] = subset_interpolated['WSPD'].interpolate() #interpolating wind speed
    subset_interpolated['timestamp_'] = subset_interpolated['timestamp_'].interpolate() #interpolating timestamp_
    interpolated_subsets.append(subset_interpolated)

# print(interpolated_subsets)

gdfw2_time_interpolated = pd.concat(interpolated_subsets)
# gdfw2_time_interpolated

# filling up the missing lats and lons

for i in range(1, num_rows):
  if (gdfw2_time_interpolated.iloc[i]['lat'] != float) and (gdfw2_time_interpolated.iloc[i]['lon'] != float):
    gdfw2_time_interpolated[['lat', 'lon']] = gdfw2_time_interpolated[['lat', 'lon']].fillna(method='ffill')
    # gdfw2_time_interpolated.iloc[i]['lat'].fillna(method='ffill')
    # gdfw2_time_interpolated.iloc[i]['lon'].fillna(method='ffill')
  else:
    continue

gdfw2_final = gdfw2_time_interpolated.round({'lat': 4, 'lon':4})

gdfw2_final

from enfactor_calc import get_weather_data # obtain the weather data at the exact location and timestamp of the ship
from enfactor_calc import get_enfactors # calculate the enfactor for the ship

# calculate own_enfactors for all own ships
num_threads = 16

own_enfactor_array = []

# Prepare the input data as lists for parallel processing
lat1_list = lat1.tolist()
lon1_list = lon1.tolist()
timestamp_list = processed_data.timestamp_.tolist()
own_course_list = processed_data.own_course.tolist()


# Create a ThreadPoolExecutor
with ThreadPoolExecutor(max_workers=num_threads) as executor:
    # Process each location in parallel
    for lat, lon, timestamp, own_course in tqdm(zip(lat1_list, lon1_list, timestamp_list, own_course_list), 
                                                   total=len(processed_data), 
                                                   desc='Calculating enfactors'):
        result = executor.submit(get_enfactors, lat, lon, timestamp, own_course, gdfw2_final, [0.28947368, 0.47368421, 0.23684211])
        own_enfactor_array.append(result.result())

print("Enfactors calculation completed.")

# calculate own_enfactors for all target ships
num_threads = 16

target_enfactor_array = []

# Prepare the input data as lists for parallel processing
lat2_list = lat2.tolist()
lon2_list = lon2.tolist()
timestamp_list = processed_data.timestamp_.tolist()
target_course_list = processed_data.target_course.tolist()


# Create a ThreadPoolExecutor
with ThreadPoolExecutor(max_workers=num_threads) as executor:
    # Process each location in parallel
    for lat, lon, timestamp, target_course in tqdm(zip(lat2_list, lon2_list, timestamp_list, target_course_list), 
                                                   total=len(processed_data), 
                                                   desc='Calculating enfactors'):
        result = executor.submit(get_enfactors, lat, lon, timestamp, target_course, gdfw2_final, [0.28947368, 0.47368421, 0.23684211])
        target_enfactor_array.append(result.result())

print("Enfactors calculation completed.")

# calculate the resultant enfactor using 75:25 weightage ratio for own ship and target ship
enfactor_array = np.array([])
enfactor_array = 0.75 * np.array(own_enfactor_array) + 0.25 * np.array(target_enfactor_array)

enfactor_df = pd.DataFrame(enfactor_array, columns=["Enfactors"])

# Save to a pickle file
enfactor_df.to_pickle('Final\\enfactor_array.pickle')
