==> 4_Pira_Consequence_Fuzzy_Logic.ipynb <==
# import all necessary libraries
import numpy as np
import pandas as pd
import importlib
import random
from tqdm import tqdm
import datetime
import geopandas as gpd
from concurrent.futures import ThreadPoolExecutor, as_completed

import numpy as np
!pip install scikit-fuzzy
import skfuzzy as fuzz
from skfuzzy import control as ctrl

# Set random seed
random.seed(42)
np.random.seed(42)

y_cri = pd.read_pickle('/content/y_cri_df.pkl')

y_cri

enfactor_df = pd.read_pickle('/content/enfactor_array.pickle')

enfactor_df

# Setting up the Fuzzy Inference System to combine the basic CRI (y_cri) with the enfactors
# Define input variables (indices)
enfactors = ctrl.Antecedent(np.arange(0, 1.0001, 0.0001), 'enfactors')
basic_risk_index = ctrl.Antecedent(np.arange(0, 1.0001, 0.0001), 'basic_risk_index')

# Define output variable (ecri_w)
ecri_w = ctrl.Consequent(np.arange(0, 1.0001, 0.0001), 'ecri_w')

# Define membership functions for the input variables

basic_risk_index['vs'] = fuzz.trimf(basic_risk_index.universe, [0, 0, 0.33])
basic_risk_index['s'] = fuzz.trimf(basic_risk_index.universe, [0, 0.33, 0.66])
basic_risk_index['m'] = fuzz.trimf(basic_risk_index.universe, [0.33, 0.66, 1])
basic_risk_index['b'] = fuzz.trimf(basic_risk_index.universe, [0.66, 1, 1])

enfactors['vs'] = fuzz.trimf(enfactors.universe, [0, 0, 0.33])
enfactors['s'] = fuzz.trimf(enfactors.universe, [0, 0.33, 0.66])
enfactors['m'] = fuzz.trimf(enfactors.universe, [0.33, 0.66, 1])
enfactors['b'] = fuzz.trimf(enfactors.universe, [0.66, 1, 1])


# Define membership functions for the output variable
ecri_w['vs'] = fuzz.trimf(ecri_w.universe, [0, 0, 0.33])
ecri_w['s'] = fuzz.trimf(ecri_w.universe, [0, 0.33, 0.66])
ecri_w['m'] = fuzz.trimf(ecri_w.universe, [0.33, 0.66, 1])
ecri_w['b'] = fuzz.trimf(ecri_w.universe, [0.66, 1, 1])


# Define fuzzy rules to combine the indices


#first and second column rules
rule1 = ctrl.Rule(basic_risk_index['vs'] & (enfactors['vs'] | enfactors['s']), ecri_w['vs'])
rule2 = ctrl.Rule(basic_risk_index['s'] & (enfactors['vs'] | enfactors['s']), ecri_w['s'])
rule3 = ctrl.Rule(basic_risk_index['m'] & (enfactors['vs'] | enfactors['s']), ecri_w['m'])
rule4 = ctrl.Rule(basic_risk_index['b'] & (enfactors['vs'] | enfactors['s']), ecri_w['b'])


#third column rules
rule5 = ctrl.Rule((basic_risk_index['vs'] | basic_risk_index['s']) & enfactors['m'], ecri_w['s'])
rule6 = ctrl.Rule(basic_risk_index['m'] & enfactors['m'], ecri_w['m'])
rule7 = ctrl.Rule(basic_risk_index['b'] & enfactors['m'], ecri_w['b'])

#fourth column rules
rule8 = ctrl.Rule(basic_risk_index['vs'] & enfactors['b'], ecri_w['s'])
rule9 = ctrl.Rule(basic_risk_index['s'] & enfactors['b'], ecri_w['m'])
rule10 = ctrl.Rule((basic_risk_index['m'] | basic_risk_index['b']) & enfactors['b'], ecri_w['b'])


# Create a control system
index_ctrl = ctrl.ControlSystem([rule1, rule2, rule3, rule4, rule5, rule6, rule7, rule8, rule9, rule10])

# Create a simulation
index_sim = ctrl.ControlSystemSimulation(index_ctrl)

ecri_w_result = []
for index1, index2 in tqdm(zip(y_cri['cri'], enfactor_df['Enfactors']), total=len(y_cri['cri'])):
  index_sim.input['basic_risk_index'] = index1
  index_sim.input['enfactors'] = index2
  index_sim.compute()
  ecri_w_result.append(index_sim.output['ecri_w'])

ecri_w_result

ecri_w_df = pd.DataFrame(ecri_w_result, columns=["ECRI-W"])

# Save to a pickle file
ecri_w_df.to_pickle('/content/ecriw_w_df.pkl')
